Refactor helpers to avoid naming conflicts and use `riscv_ext_intrinsics` for RISC-V intrinsics
