+++ crates/contracts/core/ab-contract-file/src/contract_instruction_prototype.rs
@@ -1,4 +1,5 @@
-use ab_riscv_interpreter::rv64::b::zbc::zbc_helpers;
+use ab_riscv_interpreter::rv64::b::zbb::rv64_zbb_helpers;
+use ab_riscv_interpreter::rv64::b::zbc::rv64_zbc_helpers;
 use ab_riscv_interpreter::{
     ExecutableInstruction, ExecutionError, InterpreterState, ProgramCounter,
     SystemInstructionHandler, VirtualMemory,
