+++ crates/execution/ab-riscv-interpreter/src/rv32/zk/zbkb.rs
@@ -1,5 +1,6 @@
 //! RV32 Zbkb extension
 
+pub mod rv32_zbkb_helpers;
 #[cfg(test)]
 mod tests;
 
@@ -52,30 +53,7 @@ where
                 // for i in 0..16.
                 let src = state.regs.read(rs1);
 
-                // TODO: Miri is excluded because corresponding intrinsic is not implemented there
-                let value = cfg_select! {
-                    all(not(miri), target_arch = "riscv32", target_feature = "zbkb") => {
-                        core::arch::riscv32::zip(src)
-                    }
-                    _ => {{
-                        let lo = src & 0x0000_FFFF;
-                        let hi = src >> 16;
-
-                        // Spread each 16-bit half into alternating bits.
-                        // Classic SWAR interleave for 16-bit -> 32-bit Morton.
-                        #[inline(always)]
-                        fn spread(mut x: u32) -> u32 {
-                            x = (x | (x << 8)) & 0x00FF_00FF;
-                            x = (x | (x << 4)) & 0x0F0F_0F0F;
-                            x = (x | (x << 2)) & 0x3333_3333;
-                            (x | (x << 1)) & 0x5555_5555
-                        }
-
-                        spread(lo) | (spread(hi) << 1)
-                    }}
-                };
-
-                state.regs.write(rd, value);
+                state.regs.write(rd, rv32_zbkb_helpers::zip(src));
             }
             Self::Unzip { rd, rs1 } => {
                 // Inverse of zip: gather bits of rs1 so that
@@ -84,28 +62,7 @@ where
                 // for i in 0..16.
                 let src = state.regs.read(rs1);
 
-                // TODO: Miri is excluded because corresponding intrinsic is not implemented there
-                let value = cfg_select! {
-                    all(not(miri), target_arch = "riscv32", target_feature = "zbkb") => {
-                        core::arch::riscv32::unzip(src)
-                    }
-                    _ => {{
-                        #[inline(always)]
-                        fn compact(mut x: u32) -> u32 {
-                            x &= 0x5555_5555;
-                            x = (x | (x >> 1)) & 0x3333_3333;
-                            x = (x | (x >> 2)) & 0x0F0F_0F0F;
-                            x = (x | (x >> 4)) & 0x00FF_00FF;
-                            (x | (x >> 8)) & 0x0000_FFFF
-                        }
-
-                        let lo = compact(src);
-                        let hi = compact(src >> 1);
-                        lo | (hi << 16)
-                    }}
-                };
-
-                state.regs.write(rd, value);
+                state.regs.write(rd, rv32_zbkb_helpers::unzip(src));
             }
         }
