+++ crates/execution/ab-riscv-interpreter/src/rv32/zk/zbkb/rv32_zbkb_helpers.rs
@@ -1,0 +1,53 @@
+//! Opaque helpers for RV32 Zbkb extension
+
+#[inline(always)]
+#[doc(hidden)]
+pub fn zip(src: u32) -> u32 {
+    cfg_select! {
+        all(not(miri), target_arch = "riscv32", target_feature = "zbkb") => {
+            // SAFETY: Compile-time checked for supported feature
+            unsafe { core::arch::riscv32::zip(src) }
+        }
+        _ => {{
+            let lo = src & 0x0000_FFFF;
+            let hi = src >> 16;
+
+            // Spread each 16-bit half into alternating bits.
+            // Classic SWAR interleave for 16-bit -> 32-bit Morton.
+            #[inline(always)]
+            fn spread(mut x: u32) -> u32 {
+                x = (x | (x << 8)) & 0x00FF_00FF;
+                x = (x | (x << 4)) & 0x0F0F_0F0F;
+                x = (x | (x << 2)) & 0x3333_3333;
+                (x | (x << 1)) & 0x5555_5555
+            }
+
+            spread(lo) | (spread(hi) << 1)
+        }}
+    }
+}
+
+#[inline(always)]
+#[doc(hidden)]
+pub fn unzip(src: u32) -> u32 {
+    cfg_select! {
+        all(not(miri), target_arch = "riscv32", target_feature = "zbkb") => {
+            // SAFETY: Compile-time checked for supported feature
+            unsafe { core::arch::riscv32::unzip(src) }
+        }
+        _ => {{
+            #[inline(always)]
+            fn compact(mut x: u32) -> u32 {
+                x &= 0x5555_5555;
+                x = (x | (x >> 1)) & 0x3333_3333;
+                x = (x | (x >> 2)) & 0x0F0F_0F0F;
+                x = (x | (x >> 4)) & 0x00FF_00FF;
+                (x | (x >> 8)) & 0x0000_FFFF
+            }
+
+            let lo = compact(src);
+            let hi = compact(src >> 1);
+            lo | (hi << 16)
+        }}
+    }
+}
