+++ crates/execution/ab-riscv-interpreter/src/rv64/b/zbc.rs
@@ -1,8 +1,8 @@
 //! RV64 Zbc extension
 
+pub mod rv64_zbc_helpers;
 #[cfg(test)]
 mod tests;
-pub mod zbc_helpers;
 
 use crate::{ExecutableInstruction, ExecutionError, InterpreterState};
 use ab_riscv_macros::instruction_execution;
@@ -27,57 +27,22 @@ where
     ) -> Result<ControlFlow<()>, ExecutionError<Reg::Type, CustomError>> {
         match self {
             Self::Clmul { rd, rs1, rs2 } => {
-                // Only here to prevent compiler warnings about unused `zbc_helpers` module
-                let () = zbc_helpers::PLACEHOLDER;
                 let a = state.regs.read(rs1);
                 let b = state.regs.read(rs2);
 
-                // TODO: Miri is excluded because corresponding intrinsic is not implemented there
-                let value = cfg_select! {
-                    all(not(miri), target_arch = "riscv64", target_feature = "zbkc") => {
-                        core::arch::riscv64::clmul(a as usize, b as usize) as u64
-                    }
-                    _ => {{
-                        let result = zbc_helpers::clmul_internal(a, b);
-                        result as u64
-                    }}
-                };
-
-                state.regs.write(rd, value);
+                state.regs.write(rd, rv64_zbc_helpers::clmul(a, b));
             }
             Self::Clmulh { rd, rs1, rs2 } => {
                 let a = state.regs.read(rs1);
                 let b = state.regs.read(rs2);
 
-                // TODO: Miri is excluded because corresponding intrinsic is not implemented there
-                let value = cfg_select! {
-                    all(not(miri), target_arch = "riscv64", target_feature = "zbkc") => {
-                        core::arch::riscv64::clmulh(a as usize, b as usize) as u64
-                    }
-                    _ => {{
-                        let result = zbc_helpers::clmul_internal(a, b);
-                        (result >> 64) as u64
-                    }}
-                };
-
-                state.regs.write(rd, value);
+                state.regs.write(rd, rv64_zbc_helpers::clmulh(a, b));
             }
             Self::Clmulr { rd, rs1, rs2 } => {
                 let a = state.regs.read(rs1);
                 let b = state.regs.read(rs2);
 
-                // TODO: Miri is excluded because corresponding intrinsic is not implemented there
-                let value = cfg_select! {
-                    all(not(miri), target_arch = "riscv64", target_feature = "zbc") => {
-                        core::arch::riscv64::clmulr(a as usize, b as usize) as u64
-                    }
-                    _ => {{
-                        let result = zbc_helpers::clmul_internal(a, b);
-                        (result >> 63) as u64
-                    }}
-                };
-
-                state.regs.write(rd, value);
+                state.regs.write(rd, rv64_zbc_helpers::clmulr(a, b));
             }
         }
