+++ crates/execution/ab-riscv-interpreter/src/rv64/zk/zbkc.rs
@@ -1,6 +1,6 @@
 //! RV64 Zbkc extension (subset of Zbc extension)
 
-use crate::rv64::b::zbc::zbc_helpers;
+use crate::rv64::b::zbc::rv64_zbc_helpers;
 use crate::{ExecutableInstruction, ExecutionError, InterpreterState};
 use ab_riscv_macros::instruction_execution;
 use ab_riscv_primitives::instructions::rv64::zk::zbkc::Rv64ZbkcInstruction;
